Move mempool RPCs to rpc/mempool
Can be reviewed with
--color-moved=dimmed-zebra --color-moved-ws=ignore-all-space
